switch network onClick during request-and-claim
