+++ packages/dapp/src/components/common/ConnectWeb3.jsx
@@ -1,24 +1,8 @@
-import {
-  Badge,
-  Button,
-  Flex,
-  Image,
-  Text,
-  Tooltip,
-  useToast,
-} from '@chakra-ui/react';
-import MetamaskFox from 'assets/metamask-fox.svg';
+import { Button, Flex, Text } from '@chakra-ui/react';
 import { useWeb3Context } from 'contexts/Web3Context';
 import { useBridgeDirection } from 'hooks/useBridgeDirection';
+import { useRenderChain } from 'hooks/useRenderChain';
 import { WalletFilledIcon } from 'icons/WalletFilledIcon';
-import { NON_ETH_CHAIN_IDS } from 'lib/constants';
-import {
-  getNetworkName,
-  getWalletProviderName,
-  handleWalletError,
-  logError,
-} from 'lib/helpers';
-import { addChainToMetaMask } from 'lib/metamask';
 import React, { useCallback } from 'react';
 
 export const ConnectWeb3 = () => {
@@ -27,80 +11,9 @@ export const ConnectWeb3 = () => {
     foreignChainId,
     label: bridgeLabel,
   } = useBridgeDirection();
-  const {
-    connectWeb3,
-    loading,
-    account,
-    disconnect,
-    ethersProvider,
-  } = useWeb3Context();
-  const toast = useToast();
+  const { connectWeb3, loading, account, disconnect } = useWeb3Context();
 
-  const showError = useCallback(
-    msg => {
-      if (msg) {
-        toast({
-          title: 'Error',
-          description: msg,
-          status: 'error',
-          isClosable: 'true',
-        });
-      }
-    },
-    [toast],
-  );
-
-  const addChain = useCallback(
-    async chainId => {
-      await addChainToMetaMask({ chainId }).catch(metamaskError => {
-        logError({ metamaskError });
-        handleWalletError(metamaskError, showError);
-      });
-    },
-    [showError],
-  );
-
-  const renderChain = useCallback(
-    (chainId, connect = true) => {
-      const networkName = getNetworkName(chainId);
-      const isWalletMetamask =
-        getWalletProviderName(ethersProvider) === 'metamask';
-
-      return isWalletMetamask &&
-        NON_ETH_CHAIN_IDS.includes(chainId) &&
-        connect ? (
-        <Tooltip label={`Click to switch to ${networkName}`} position="auto">
-          <Badge
-            display="inline-flex"
-            alignItems="center"
-            py={1}
-            px={2}
-            m={1}
-            borderRadius={5}
-            size="1"
-            cursor="pointer"
-            colorScheme="blue"
-            onClick={() => addChain(chainId)}
-          >
-            <Image boxSize={5} src={MetamaskFox} mr={2} />
-            {networkName}
-          </Badge>
-        </Tooltip>
-      ) : (
-        <Text
-          as="span"
-          fontWeight="bold"
-          textTransform="uppercase"
-          fontSize="0.9rem"
-          color="black"
-        >
-          {' '}
-          {networkName}
-        </Text>
-      );
-    },
-    [addChain, ethersProvider],
-  );
+  const renderChain = useRenderChain();
 
   const renderBridgeLabel = useCallback(
     () => (
@@ -136,7 +49,7 @@ export const ConnectWeb3 = () => {
       w="calc(100% - 2rem)"
       mt="5rem"
       p="2rem"
-      maxW="27.5rem"
+      maxW="31rem"
       mx={4}
     >
       <Flex
