+++ packages/dapp/src/lib/metamask.js
@@ -1,29 +1,50 @@
 import { utils } from 'ethers';
-
 import {
   getExplorerUrl,
   getNetworkCurrency,
   getNetworkName,
   getRPCUrl,
-} from './helpers';
+  logError,
+} from 'lib/helpers';
 
 export const addChainToMetaMask = async ethereumChain => {
   const { chainId } = ethereumChain;
   const { name, symbol } = getNetworkCurrency(chainId);
-  return window.ethereum.request({
-    method: 'wallet_addEthereumChain',
-    params: [
-      {
-        chainId: utils.hexValue(chainId),
-        chainName: getNetworkName(chainId),
-        nativeCurrency: {
-          name,
-          symbol,
-          decimals: 18,
+
+  try {
+    await window.ethereum.request({
+      method: 'wallet_switchEthereumChain',
+      params: [
+        {
+          chainId: utils.hexValue(chainId),
         },
-        rpcUrls: [getRPCUrl(chainId)],
-        blockExplorerUrls: [getExplorerUrl(chainId)],
-      },
-    ],
-  });
+      ],
+    });
+  } catch (switchError) {
+    // This error code indicates that the chain has not been added to MetaMask.
+    if (switchError.code === 4902) {
+      try {
+        await window.ethereum.request({
+          method: 'wallet_addEthereumChain',
+          params: [
+            {
+              chainId: utils.hexValue(chainId),
+              chainName: getNetworkName(chainId),
+              nativeCurrency: {
+                name,
+                symbol,
+                decimals: 18,
+              },
+              rpcUrls: [getRPCUrl(chainId)],
+              blockExplorerUrls: [getExplorerUrl(chainId)],
+            },
+          ],
+        });
+      } catch (addError) {
+        logError({ addError });
+      }
+    } else {
+      logError({ switchError });
+    }
+  }
 };
